Refactor simulation configuration and traffic light controller
- Updated default tick rate from 10 to 5 in simulation configuration tests.
- Adjusted tick interval calculations to reflect new tick rate.
- Enhanced vehicle spawner tests to ensure proper spawning behavior with new graph connections.
- Introduced NeighborInfo data class for vehicle leader information in traffic physics.
- Implemented dynamic rerouting logic for vehicles affected by new blocks or restrictions.
- Added traffic light detection logic to manage vehicle behavior at traffic signals.
- Created comprehensive tests for traffic light controller functionality, including phase transitions and overrides.
